add timeout to communication with Modbus device
This adds a timeout variant of read_frame that is used in communication
with the Modbus device. It helps with picky Modbus devices or a bad
connection to the Modbus device to abort the read operation after a
specific time and properly shutdown the connection(s). The default
timeout is currently set to 5 seconds.

This seems to solve a cascading problem where the connection to the
Modbus device breaks, which causes several async tasks to block waiting
for something to read from the Modbus device. With some clients, which
then try again and again with new TCP connections, this leads to an
increase of open sockets, causing other errors (like 'No file
descriptors available') after some time.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>